requirements.txt
# Core Dependencies
openai>=1.0.0
chainlit>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# AI and ML
sentence-transformers>=2.0.0
chromadb>=0.4.0

# Utilities
python-dotenv>=1.0.0
keyring>=24.0.0
structlog>=23.0.0
rich>=13.0.0

# Database and Storage
sqlalchemy>=2.0.0
alembic>=1.0.0

# Image Processing
pillow>=10.0.0
requests>=2.31.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: Vertex AI (Google Cloud)
google-cloud-aiplatform>=1.35.0

# Optional: Additional Image Generation APIs
replicate>=0.20.0
stability-sdk>=0.8.0
